+++ app/src/main/java/com/example/relay/timetable/view/ScheduleBottomSheetFragment.kt
@@ -1,4 +1,4 @@
-package com.example.relay.timetable
+package com.example.relay.timetable.view
 
 import android.content.Context
 import android.os.Bundle
