+++ app/src/main/java/com/example/relay/timetable/view/TimetableEditFragment.kt
@@ -1,4 +1,4 @@
-package com.example.relay.timetable
+package com.example.relay.timetable.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,9 +8,11 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.relay.R
 import com.example.relay.databinding.FragmentTimetableEditBinding
+import com.example.relay.timetable.ScheduleRvAdapter
+import com.example.relay.timetable.TimetableInterface
+import com.example.relay.timetable.TimetableService
 import com.example.relay.timetable.models.GroupTimetableRes
 import com.example.relay.timetable.models.Schedule
-import java.sql.Time
 
 class TimetableEditFragment : Fragment(), TimetableInterface {
     private var viewBinding : FragmentTimetableEditBinding? = null
