+++ app/src/main/res/layout/fragment_schedule_bottom_sheet.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="200dp"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".timetable.ScheduleBottomSheetFragment">
+    tools:context=".timetable.view.ScheduleBottomSheetFragment">
 
     <Button
         android:id="@+id/btn_save"
